% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{pm_add_quantities}
\alias{pm_add_quantities}
\title{Adds units and measurement errors to a table extracted from the pmird database}
\usage{
pm_add_quantities(x)
}
\arguments{
\item{x}{A data frame with attributes which were extracted from the pmird
database.}
}
\value{
\code{x} with added measurement units and uncertainties.
}
\description{
Adds units and measurement errors to a table extracted from the pmird database
}
