% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{pm_append_rows_csv}
\alias{pm_append_rows_csv}
\alias{pm_append_rows_csv_persons}
\alias{pm_append_rows_csv_attributes}
\title{Appending new entries in a table to a csv file with similar columns}
\usage{
pm_append_rows_csv_persons(con, file)

pm_append_rows_csv_attributes(con, file)
}
\arguments{
\item{con}{A connection to the pmird database.}

\item{file}{A character value with the path to the csv file to which to
append rows.}
}
\value{
\code{TRUE} if appending rows was successful, otherwise \code{FALSE}.
}
\description{
Helper function to append new entries in a pmird database table (indicated in
the function name) to a csv file with similar columns.
}
\keyword{Internal}
