% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pm_dm_add_keys.R
\name{pm_dm_add_keys}
\alias{pm_dm_add_keys}
\title{Adds primary and foreign keys to a \code{dm} object with tables from the pmird database}
\usage{
pm_dm_add_keys(x, dm_pmird, progress = NA)
}
\arguments{
\item{x}{A \code{\link[dm]{dm}} object with tables from pmird database.}

\item{dm_pmird}{A \code{\link[dm]{dm}} object representing the pmird
database.}

\item{progress}{A logical value. See \code{\link[dm]{dm_examine_constraints}}.}
}
\value{
\code{x} with added primary and foreign keys.
}
\description{
\code{pm_dm_add_keys} is a helper function that takes a \code{dm} object
with tables from the pmird database, adds the primary and foreign keys,
and checks whether these are valid, using \code{\link[dm]{dm_examine_constraints}}.
}
