% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pm_get_citations.R
\name{pm_get_citations}
\alias{pm_get_citations}
\title{Lists datasources to cite when using data from the pmird database}
\usage{
pm_get_citations(con, x, file = NULL)
}
\arguments{
\item{con}{A connection to the pmird database. This should be an object of
class \code{\link[RMariaDB::`MariaDBConnection-class`]{MariaDBConnection}}, created
using \code{\link[RMariaDB::`MariaDBConnection-class`]{dbConnect}}.}

\item{x}{A numeric vector with values from `id_measurement` from table
`measurements` in the 'pmird' database.}

\item{file}{A character value defining the path to the bibtex file where to
write the collected bibliogrpahic information. Can be set to `NULL` to omit
exporting anything to a file.}
}
\value{
A data frame as created by `bib2df::bib2df()` with two additional
columns: `id_dataset` (the ID for the dataset) and `id_measurement` (unique
values of `x` identifying the measurement corresponding to the respective
dataset). As a side effect, unique values of column `bib` will be written
to the file specified in `file` if `! is.null(file)`.
}
\description{
`pm_get_citations()` takes values from column `id_measurement` in table
`measurements` from the pmird database and collects bibliographic
information on the data sources which have to be cited when the respective
data are used. The collected bibliographic information are exported to a
bibtex file.
}
