% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pm_get_dm.R
\name{pm_get_dm}
\alias{pm_get_dm}
\title{Access the pmird database as a \code{dm} object.}
\usage{
pm_get_dm(con, learn_keys = FALSE)
}
\arguments{
\item{con}{A connection to the pmird database. This should be an object of
class \code{\link[RMariaDB::`MariaDBConnection-class`]{MariaDBConnection}}, created
using \code{\link[RMariaDB::`MariaDBConnection-class`]{dbConnect}}.}

\item{learn_keys}{A logical value indicating if primary and foreign keys
should be assigned (\code{TRUE}) or not (\code{FALSE}).}
}
\value{
A \code{dm} object with the pmird database and all primary and
foreign keys.
}
\description{
\code{pm_get_dm} extracts tables and relations in the form of a
\code{\link[dm:dm]{dm}} object. It is a wrapper around
\code{\link[dm:dm_from_src]{dm_from_src}} which manually adds primary and
foreign keys to the tables which dm currently does not support for MariaDB
databases.
}
\examples{
\dontrun{
# connect to database
con <-
  RMariaDB::dbConnect(
    drv = RMariaDB::MariaDB(),
    dbname = "pmird",
    default.file = "~/my.cnf"
  )

# get pmird database as dm object
dm_pmird <- pm_get_dm(con, learn_keys = TRUE)

# when finished, disconnect
RMariaDB::dbDisconnect(con)
}
}
