% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pm_get_max.R
\name{pm_get_max}
\alias{pm_get_max}
\title{Takes a \code{zoomed_dm} object or a data frame and returns the largest element of a certain column.}
\usage{
pm_get_max(x, .col)
}
\arguments{
\item{x}{A \code{zoomed_dm} object or a data frame.}

\item{.col}{The name of a column in \code{x} for which to extract the largest
number.}
}
\value{
The maximum value in the column \code{.col} in \code{x} or 0 if
\code{x} has no rows.
}
\description{
\code{pm_get_max} takes a \code{zoomed_dm} object or a data frame and the
name of a specific column and returns the maximum value of this column in the
\code{zoomed_dm} object or data frame. \code{pm_get_max} is used as helper
function to extract the last identifier in a table in the pmird database.
}
\examples{
\dontrun{
# connect to database
con <-
  RMariaDB::dbConnect( # ---todo: change this
    drv = RMariaDB::MariaDB(),
    dbname = "pmird",
    default.file = "~/my.cnf"
  )

# get pmird database as dm object
dm_pmird <- pm_get_dm(con, learn_keys = TRUE)

# extract the last ID from a table
dm::dm_zoom_to(datasets) \%>\%
  pm_get_max(.col = id_dataset)

# when finished, disconnect
RMariaDB::dbDisconnect(con)
}
}
