% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pm_update_import.R
\name{pm_get_present}
\alias{pm_get_present}
\title{Provides a unique ID for entities already present in the pmird database.}
\usage{
pm_get_present(con, table_name)
}
\arguments{
\item{con}{Either a database connection (see
\code{\link[pm_update_import]{pm_update_import}}) or a data frame wit the
same format as the table indicated by \code{table_name} in the pmird database
(same column names, no additional columns).}

\item{table_name}{See \code{\link[pm_update_import]{pm_update_import}}.}
}
\value{
A data frame with the same rows as the table \code{table_name} in the
pmird database (if \code{con} is a database connection) or in the provided
data frame (id \code{x} is a data frame) and with the folowing two columns:
\enumerate{
 \item{The ID (primary key).}
 \item{A column \code{x_present} with the unique ID which can be used to
 match entities across data frames.}
}
}
\description{
\code{pm_get_present} is a helper function to define a unique ID for entities
already present in a table in the pmird database. This function is used by
\code{\link[pm_update_import]{pm_update_import}}.
}
\keyword{Internal}
