% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pm_ir_extract_peak.R
\name{pm_ir_extract_peak}
\alias{pm_ir_extract_peak}
\title{Extracts a peak from an infrared spectrum}
\usage{
pm_ir_extract_peak(x, range, peak_max = 2362, range_offset = 4)
}
\arguments{
\item{x}{An object of class \code{\link[ir:ir_new_ir]{ir}}.}

\item{range}{A data frame with one row. See
\code{\link[ir:ir_clip]{ir_clip}}. This is the range which is extracted from
\code{x}.}

\item{peak_max}{A numeric value representing the wavenumber value where the
peak maximum (or minimum if the peak has a negative direction) is.}

\item{range_offset}{A positive numeric value representing the offset which is
additionally clipped from the x axis range given by \code{range} to avoid
artifacts during baseline correction.}
}
\value{
\code{x} clipped to \code{range} (less \code{range_offset}), baseline
corrected with \code{\link[ir:ir_bc_rubberband]{ir_bc_rubberband}} and with
a new column \code{peak_direction} with value \code{"up"} if the peak is a
maximum or \code{"down"} if not.
}
\description{
\code{pm_ir_extract_peak} is a function specifically designed for pmird to
help quality assessment of mid infrared spectra. It clips a mid infrared
spectra to a defined region, baseline corrects this region, and identifies
if the peak has a positive direction or a negative direction.
}
