% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmird-package.R
\docType{package}
\name{pmird-package}
\alias{pmird}
\alias{pmird-package}
\title{pmird: R interface to the peatland mid infrared spectra database}
\description{
The 'pmird' package provides functions to interact with the 
    'pmird' database. The 'pmird' database stores mid infrared spectra of peat, 
    peat forming vegetation and dissolved organic matter along with metadata on 
    samples and additional data (e.g., elemental contents).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/henningte/pmird}
  \item Report bugs at \url{https://github.com/henningte/pmird/issues}
}

}
\author{
\strong{Maintainer}: Henning Teickner \email{henning.teickner@uni-muenster.de} (0000-0002-3993-1182)

}
\keyword{internal}
